Move reply state cleanup to RSC module
