+++ src/aggregate/aggregate_request.c
@@ -1585,31 +1585,6 @@ int AREQ_ApplyContext(AREQ *req, RedisSearchCtx *sctx, QueryError *status) {
   return REDISMODULE_OK;
 }
 
-void ChunkReplyState_Destroy(ChunkReplyState *state) {
-  // Free any stored results that weren't consumed
-  // (e.g., if timeout occurred before reply_callback ran)
-  if (state->results) {
-    for (size_t i = 0; i < array_len(state->results); i++) {
-      SearchResult_Destroy(state->results[i]);
-      rm_free(state->results[i]);
-    }
-    array_free(state->results);
-    state->results = NULL;
-  }
-
-  // Timeout edge case: cursor wasn't handled by reply_callback.
-  // See ChunkReplyState ownership model in aggregate.h for full explanation.
-  // We must clear query before Cursor_Free to prevent the AREQ_DecrRef loop.
-  if (state->cursor) {
-    state->cursor->query = NULL;
-    Cursor_Free(state->cursor);
-    state->cursor = NULL;
-  }
-
-  // Clear stored error state
-  QueryError_ClearError(&state->err);
-}
-
 void AREQ_Free(AREQ *req) {
   // Check if rootiter exists but pipeline was never built (no result processors)
   // In this case, we need to free the rootiter manually since no RPQueryIterator
